# Lvardt cell CvMembList.ml.size() should be the number of contiguous regions of the fixed step thread Memb_list