--- a/src/nrncvode/netcvode.cpp
+++ b/src/nrncvode/netcvode.cpp
@@ -1615,6 +1615,9 @@ bool NetCvode::init_global() {
                 }
             }
 
+            // Modified to also count the nodes and set the offsets for
+            // each CvMembList.ml[contig_region]
+            // See the long comment after this loop.
             for (NrnThreadMembList* tml = _nt->tml; tml; tml = tml->next) {
                 i = tml->index;
                 const Memb_func& mf = memb_func[i];
@@ -1625,42 +1628,77 @@ bool NetCvode::init_global() {
                     // singly linked list built below
                     int j;
                     for (j = 0; j < ml->nodecount; ++j) {
+                        auto offset = ml->get_storage_offset() + j;
+                        // for each Memb_list instance constructed, keep
+                        // track of its initial storage offset (i.e. offset)
+
                         int inode = ml->nodelist[j]->v_node_index;
+                        int icell = cellnum[inode];
                         Cvode& cv = d.lcv_[cellnum[inode]];
                         CvodeThreadData& z = cv.ctd_[0];
-                        if (!z.cv_memb_list_) {
+
+                        // Circumstances for creating a new CvMembList
+                        // or (due to non-contiguity of a cell), 
+                        // appending a Memb_list instance to cml->ml
+                        if (!z.cv_memb_list_) { // initialize the first
                             cml = new CvMembList{i};
-                            cml->next = nullptr;
-                            assert(cml->ml.size() == 1);
-                            cml->ml[0].nodecount = 0;
                             z.cv_memb_list_ = cml;
+                            cml->next = nullptr;
                             last[cellnum[inode]] = cml;
-                        }
-                        if (last[cellnum[inode]]->index == i) {
-                            assert(last[cellnum[inode]]->ml.size() == 1);
-                            ++last[cellnum[inode]]->ml[0].nodecount;
-                        } else {
+                            assert(cml->ml.size() == 1);
+                            assert(cml->ml[0].nodecount == 0);
+                        } else if (last[cellnum[inode]]->index != i) { // initialize next
                             cml = new CvMembList{i};
                             last[cellnum[inode]]->next = cml;
                             cml->next = nullptr;
                             last[cellnum[inode]] = cml;
                             assert(cml->ml.size() == 1);
-                            cml->ml[0].nodecount = 1;
+                            assert(cml->ml[0].nodecount == 0);
+                        } else { // if non-contiguous, append Memb_list
+                            cml = last[cellnum[inode]];
+                            auto& cvml = cml->ml.back();
+                            auto cvml_offset = cvml.get_storage_offset() + cvml.nodecount;
+                            if (cvml_offset != offset) {
+                                // not contiguous, add another Memb_list
+                                // instance to cml->ml
+                                cml->ml.emplace_back(cml->index);
+                                assert(cml->ml.back().nodecount == 0);
+                            }
                         }
+
+                        auto& cvml = cml->ml.back();
+                        if (cvml.nodecount == 0) { //first time for this Memb_List
+                            cvml.set_storage_offset(offset);
+                        }
+                        // Increment count of last Memb_list in cml->ml.
+                        ++cvml.nodecount;
                     }
                 }
             }
-            // allocate and re-initialize count
+            // each NetCvode.p[id].lcv_[nt->ncell].ctd_[0].cv_memb_list_->ml[0].nodecount
+            // is setup and ml.size() == 1.
+
+            // if ml[0] data is not contiguous, need to count number
+            // of contiguous data regions and reserve that number of Memb_list
+            // in the ml vector. The sum of the ml[i].nodecount must equal
+            // the present ml[0].nodecount and each ml[i] data must be
+            // contiguous.
+            // Ideally the node permutation would be such that each cell
+            // is contiguous. That is almost the case with the default
+            // permutation. But the cell root nodes are all at the
+            // beginning, and thereafter the rest of the cell is contiguous.
+            // For all density mechanims (except extracellular) this is fine
+            // since those mechanisms do not exist at the root node. And
+            // most of the time, there are no POINT_PROCESSes in the
+            // root nodes. So, we expect that most of the time, we have
+            // the default node permutation and that will result in
+            // almost all CvMembList.ml.size() == 1 with a very occasional
+            // size() == 2 
+
+
             std::vector<CvMembList*> cvml(d.nlcv_);
             for (i = 0; i < d.nlcv_; ++i) {
-                cvml[i] = d.lcv_[i].ctd_[0].cv_memb_list_;
-                for (cml = cvml[i]; cml; cml = cml->next) {
-                    // non-contiguous mode, so we're going to create a lot of 1-element Memb_list
-                    // inside cml->ml
-                    cml->ml.reserve(cml->ml[0].nodecount);
-                    // remove the single entry from contiguous mode
-                    cml->ml.clear();
-                }
+                cvml[i] = d.lcv_[i].ctd_[0].cv_memb_list_; // whole cell in thread
             }
             // fill pointers (and nodecount)
             // now list order is from 0 to n_memb_func
@@ -1674,20 +1712,32 @@ bool NetCvode::init_global() {
                         int icell = cellnum[ml->nodelist[j]->v_node_index];
                         if (cvml[icell]->index != i) {
                             cvml[icell] = cvml[icell]->next;
-                            assert(cvml[icell] && cvml[icell]->index);
+                            assert(cvml[icell] && cvml[icell]->index == i);
                         }
-                        cml = cvml[icell];
-                        auto& newml = cml->ml.emplace_back(cml->index /* mechanism type */);
-                        newml.nodecount = 1;
-                        newml.nodelist = new Node*[1];
-                        newml.nodelist[0] = ml->nodelist[j];
-                        newml.nodeindices = new int[1]{ml->nodeindices[j]};
-                        newml.prop = new Prop* [1] { ml->prop[j] };
-                        if (!mf.hoc_mech) {
-                            newml.set_storage_offset(ml->get_storage_offset() + j);
-                            newml.pdata = new Datum* [1] { ml->pdata[j] };
+                        auto& cml = cvml[icell];
+                        int kk = j;
+                        for (auto& newml: cml->ml) {
+                            auto nodecount = newml.nodecount;
+                            if (! newml.nodelist) {
+                                newml.nodelist = new Node*[nodecount];
+                                newml.nodeindices = new int[nodecount];
+                                newml.prop = new Prop* [nodecount];
+                                if (!mf.hoc_mech) {
+                                    newml.pdata = new Datum* [nodecount];
+                                }
+                                for (int k = 0; k < nodecount; ++k) {
+                                    newml.nodelist[k] = ml->nodelist[kk+k];
+                                    newml.nodeindices[k] = ml->nodeindices[kk+k];
+                                    assert (cellnum[newml.nodeindices[k]] == cellnum[ml->nodeindices[j]]);
+                                    newml.prop[k] = ml->prop[kk+k];
+                                    if (!mf.hoc_mech) {
+                                        newml.pdata[k] = ml->pdata[kk+k];
+                                    }
+                                }
+                                kk += nodecount;
+                                newml._thread = ml->_thread;
+                            }
                         }
-                        newml._thread = ml->_thread;
                     }
                 }
             }