There are serious performance problems with the local variable time step method in comparison with 8.2 This PR is a prerequisite for fixing one of them.

The 9.0  implementation for local variable time step (each cell is integrated by its own cvode instance) demanded that a cell's ```CvMembList.ml.size()``` could be either 1 or ```Memb_list .nodecount```.
Thus, if there were multiple cells in a thread, the ```ml.size() == nodecount```, and each ```Memb_list``` instance was limited to a single Node*.

A subsequent PR will default sort the fixed step nodes so that Nodes of each cell will be as contiguous as possible. Presently, the default is that only nodes in a Section are contiguous.