@nrnhines 2025-01-07
> From a high level perspective, do you happen to have some performance numbers related to this change, or do we also need #3300 to be able to see a (substantial) difference? Also, are the improvements consistent on both Linux and MacOS?

Ironically, this is performant equivalent to #3300 for a single cell model. (For a single cell model global vardt and local vardt are conceptually equivalent since in both cases there is only a single instance of the cvode integrator). I say "ironically" because the my first sight of the issue was for the  single cell modeldb model 267666 discrepancy between master local and global cv:advance_tn time and instructions retired:
 https://github.com/neuronsimulator/nrn/issues/3283#issuecomment-2564929982
```
Task Type / Function / Call Stack	CPU Time	Instructions Retired	Task Time	Task Count
cv:advance_tn	14.851s	177,837,700,000	15.277s	207,743   #local vardt
cv:advance_tn   6.207s  67,917,900,000  6.354s  207,743   #global vardt
```
The reason this PR is not an endpoint solution to the problem is that default node ordering is globally section depth first which interleaves sections of different cells. Here is a ringtest sim time comparison between master, this pr, and #3300 using the intel compiler
```
$ nrniv -python ringtest.py -rparm -nring 32 -npt 1 -compart 1 3 -method 2 -2nd_order_thresh
```
```
version simtime (s)
master  6.7
this PR 3.8
PR3300  2.8
```
I have not checked the mac performance.

> Regarding the CI failure on MacOS, I don't really see how it's related to this change (something about fmt? But the PR doesn't touch those parts at all), so we can probably ignore it.

I believe that was localized to the new NMODL and I didn't look closely about where the specific failing CI's were getting that version. I do notice that locally my PR and the master are both getting 49fdfe6c from https://github.com/BlueBrain/nmodl