@nrnhines 2025-05-28
@JCGoran I'm ready to start pushing through this series of PR's of which this is the first step. At the end of the series the 9.0 performance will equal or exceed 8.2. I want to keep the PR's localized to specific cvode submodules in order to avoid a mega PR that would be difficult to understand as a squashed blob.